pixelate and using alphamerge
making a pixelated image and using alphamerge and a black and white image or video as a mask. The white areas are the areas that the image you select "sticks" to and then all the black area will be made transparent and then you can overlay what you have created on top of another video or image.